+++ scripts/macros/2014/classFeatures/warlock/undead/graveTouched.js
@@ -1,27 +1,22 @@
-import {combatUtils, constants, dialogUtils, effectUtils, genericUtils, itemUtils, rollUtils, workflowUtils} from '../../../../../utils.js';
-
+import {combatUtils, dialogUtils, effectUtils, genericUtils, rollUtils, workflowUtils} from '../../../../../utils.js';
 async function damage({trigger: {entity: item}, workflow}) {
-    if (workflow.hitTargets.size !== 1) return;
+    if (!workflow.hitTargets.size) return;
     if (!workflowUtils.isAttackType(workflow, 'attack')) return;
-    if (!combatUtils.perTurnCheck(item, 'graveTouched', true, workflow.token.id)) return;
+    if (!item.system.uses.value) return;
+    if (!combatUtils.isOwnTurn(workflow.token)) return;
     let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: item.name}));
     if (!selection) return;
-    await workflowUtils.completeItemUse(item);
-    await combatUtils.setTurnCheck(item, 'graveTouched');
-    let newRolls = [];
-    for (let roll of workflow.damageRolls) {
-        newRolls.push(rollUtils.getChangedDamageRoll(roll, 'necrotic'));
-    }
-    await workflow.setDamageRolls(newRolls);
-    let formOfDread = effectUtils.getEffectByIdentifier(workflow.actor, 'formOfDread');
-    if (formOfDread) {
-        let bonusFormula = '1d' + workflow.damageRoll.dice[0].faces + '[necrotic]';
-        await workflowUtils.bonusDamage(workflow, bonusFormula, {damageType: 'necrotic'});
-    }
+    await workflowUtils.syntheticItemRoll(item, [workflow.token], {consumeResources: true, consumeUsage: true});
+    workflow.damageRolls = workflow.damageRolls.map(roll => {
+        return rollUtils.getChangedDamageRoll(roll, 'necrotic');
+    });
+    await workflow.setDamageRolls(workflow.damageRolls);
+    let formOfDread = effectUtils.getEffectByIdentifier(workflow.actor, 'formOfDreadActive');
+    if (formOfDread) await workflowUtils.bonusDamage(workflow, '1d' + workflow.damageRolls[0].dice[0].faces, {damageType: 'necrotic'});
 }
 export let graveTouched = {
     name: 'Grave Touched',
-    version: '1.1.0',
+    version: '1.4.23',
     midi: {
         actor: [
             {